refactor: clean docs and move code
